application-suptypes
This patch adds the concept of 'subtypes' for applications.

When a CDR is recorded, it's Sip-Application-Type may contain a subtype
(for example, audio.incoming).

When the rating engine looks for a matching billing rate, it will match
on the full application.subtype. This allows billing rates to be more
specific, for example recording different rates for incoming traffic
versus outgoing.